Generate only one twins tuple for each distinct pair
Previously we would have tuples like ('A1', 'B1', x) and ('B1', 'A1',
x), which is a bit wasteful since we are needlessly performing
elimination over the same pair of twins twice. This changes twins
list generation to enforce that a < b for tuple (a, b, x) so we only get
one tuple per pair.